flask/requirements.txt
Flask>=3.0.2
Werkzeug>=3.0.1
tensorflow==2.15.0
Pillow>=10.2.0
flask-cors>=4.0.0
numpy>=1.26.4
